Added warning in README regarding current instability of generate.sh
generate.sh suffers from buffering, synchronisation and race condition
problems when run at scale (i.e., the scale doesn't increase the
probability of these things happening, just the expectation of such), as
it tries to do everything through Unix pipelines. This includes, but is
probably not limited to:

* Interleaved output from multiple processes appending to the same file
  at once.
* Temporary files being deleted before a process is finished with them.
* Missing output on standard out/error due to complicated pipelines.

Some synchronisation efforts have been put in place (e.g., a semaphore
system for the main aggregation by filetype), but it's not enough to get
around this fragility.

Options would be:

1. Redouble efforts to synchronise processes. This would make the script
   increasingly complex.
2. Move away from the streaming approach and just dump intermediate
   results to disk. This would make the process a whole lot slower and
   more expensive, in terms of resource usage.
3. Use something other than Bash. I'm not sure that this is viable (what
   would the alternative be? C?)